% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.structure.activities.R
\name{categorize.structure.activities}
\alias{categorize.structure.activities}
\title{categorize.structure.activities}
\usage{
categorize.structure.activities(datadir, date)
}
\arguments{
\item{datadir}{A string representing the directory where the activity data files are located.}

\item{date}{A string representing the date to be used in the filename for loading and saving data.}
}
\value{
The function does not return a value but saves two CSV files with the processed data.
}
\description{
'categorize.structure.activities' loads activity data from the MLM app (as structured by 'link.app.castor'), categorizes the activities, calculates the duration and frequency of each activity, and saves the processed data.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads activity data from a CSV file.
\item Calculates the total duration of each activity.
\item Categorizes activities into 24-hour movement behaviors (e.g., SB for sedentary behavior, PA for physical activity, and sleep)
\item Saves the categorized data with durations into a new CSV file.
\item Converts the long format activity data into a wide format, summarizing daily activity duration and frequency.
\item Adds day of the week and weekend information.
\item Saves the structured data into another CSV file.
}
}
