% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.save.plots.R
\name{combine.save.plots}
\alias{combine.save.plots}
\title{combine.save.plots}
\usage{
combine.save.plots(plot1, plot2, format, savedir, filename)
}
\arguments{
\item{plot1}{Plot object 1}

\item{plot2}{Plot object 2}

\item{format}{A string that contains the placement of the acceleromter. One of c("boxplots", "overlay")}

\item{savedir}{A string that indicates the path to which the plots will be saved}

\item{filename}{A string that indicates the name of the file and ends with the extension (e.g. .png, .jpeg)}
}
\value{
DESCRIPTION HERE
}
\description{
'combine.save.plots' loads the cleaned app data
}
