% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.boxplot.R
\name{create.boxplot}
\alias{create.boxplot}
\title{create.boxplot}
\usage{
create.boxplot(data_long, metric, per, order_categories)
}
\arguments{
\item{data_long}{Data description here}

\item{metric}{A string that contains the name of the metric. One of c("ENMO", "MAD")}

\item{per}{A string that indicating the plot windows. One of c("activity", "behavior, "posture")}

\item{order_categories}{Vector with the label names in the desired order}
}
\value{
DESCRIPTION HERE
}
\description{
'create.boxplot' function description
}
