% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.overlayplot.R
\name{create.overlayplot}
\alias{create.overlayplot}
\title{create.overlayplot}
\usage{
create.overlayplot(data.app, filepath.hip, filepath.wrist, pp, savedir)
}
\arguments{
\item{data.app}{A data frame containing the app data, which includes columns for \code{castorID}, \code{date}, \code{startTime}, \code{endTime}, and \code{activity}.}

\item{filepath.hip}{The file path to the directory containing the cleaned hip accelerometer \code{.cwa.RData} files.}

\item{filepath.wrist}{The file path to the directory containing the wrist accelerometer \code{.cwa.RData} files.}

\item{pp}{The participant identifier (castorID) to be processed.}

\item{savedir}{The directory where the plots will be saved.}
}
\value{
This function does not return a value. It saves plots as JPEG files.
}
\description{
'create.overlayplot' creates overlay plots of activity data recorded via the MLM app and accelerometer data (hip and wrist worn). The plots are saved as JPEG files.
}
\details{
The function first filters the app data for the specified participant. It then loads the corresponding accelerometer data for the hip and wrist, if available. For each unique day of app data, it generates overlay plots showing the app-recorded activities and the accelerometer-derived ENMO values. The plots are saved to the specified directory.
}
\examples{
\dontrun{
create.overlayplot(data.app = app_data, filepath.hip = "path/to/hip/data", 
                   filepath.wrist = "path/to/wrist/data", pp = "participant_01", 
                   savedir = "path/to/save/plots")
}

}
