% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.axivity.R
\name{estimates.axivity}
\alias{estimates.axivity}
\title{estimates.axivity}
\usage{
estimates.axivity(validdatadir)
}
\arguments{
\item{validdatadir}{Path to the directory where the preprocessed valid data is stored.}
}
\value{
A data.frame object with calculated metrics.
}
\description{
'estimates.axivity' calculates various metrics from the valid days' data and returns the outcomes, including
Median acceleration, 25th percentile, and additional calculated metrics for all data files.
}
\details{
This function calculates metrics such as ENMO, MAD, intensity gradient (ig), various percentiles, and
additional information like participant ID, date, day of the week, etc. The results are returned in a
structured data.frame.
}
\examples{
# Example usage:
data_dir <- "/path/to/valid_data"
result <- estimates.axivity(data_dir)
}
