% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.app.castor.R
\name{link.app.castor}
\alias{link.app.castor}
\title{link.app.castor}
\usage{
link.app.castor(datadir.app, datadir.castor, filename.castor, date)
}
\arguments{
\item{datadir.app}{A character string specifying the directory where the MLM app data files are located.}

\item{datadir.castor}{A character string specifying the directory where the Castor data file is located.}

\item{filename.castor}{A character string specifying the name of the Castor data file.}

\item{date}{A character string specifying the date used for naming the input and output CSV files.}
}
\value{
This function does not return a value. It saves two CSV files with the linked data in the specified directory.
}
\description{
'link.app.castor' links Castor IDs and cohort numbers to motor milestones and activities data collected from the MLM app. The linked data is then saved as CSV files.
}
