% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.app.R
\name{load.app}
\alias{load.app}
\title{load.app}
\usage{
load.app(filepath, filename, cohort, measurementperiod, sep)
}
\arguments{
\item{filepath}{A string that contains the path to the app data}

\item{filename}{A string that contains the name of the Excel app data file}

\item{cohort}{Vector indicating the cohort for the data selection. c(1, 2, 3)}

\item{measurementperiod}{Integer indicating the measurement period for the data selection. One of c(1, 2, 3)}
}
\value{
DESCRIPTION HERE
}
\description{
'load.app' loads the cleaned app data
}
