% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.castor.R
\name{load.castor}
\alias{load.castor}
\title{load.castor}
\usage{
load.castor(filepath, filename, cohort)
}
\arguments{
\item{filepath}{A string that contains the path to the data}

\item{filename}{A string that contains the name of the Excel app data file}

\item{cohort}{Vector indicating the cohort for the data selection c(1, 2, 3)}
}
\value{
DESCRIPTION HERE
}
\description{
'load.castor' loads the study data retrieved from Castor EDS
}
