% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.structure.axivity.R
\name{load.structure.axivity}
\alias{load.structure.axivity}
\title{load.structure.axivity}
\usage{
load.structure.axivity(
  tz = "Europe/Amsterdam",
  filepath,
  outputdir,
  validhours = 12,
  epochlength = 5,
  processeddir,
  overwrit = FALSE
)
}
\arguments{
\item{tz}{Time zone specification to be used, default is "Europe/Amsterdam".}

\item{filepath}{Path to the root of the accelerometer data.}

\item{outputdir}{Path to the root directory in which all milestone data will be saved.}

\item{validhours}{Integer indicating the number of hours for a day to be considered valid (midnight-midnight), default is 12.}

\item{epochlength}{Integer indicating the epoch length in seconds, default is 5.}

\item{processeddir}{Path to the directory in which the generated milestone data is saved.}

\item{overwrit}{Logical indicating whether to overwrite existing data, default is FALSE.}
}
\description{
'load.structure.axivity' Loads the files using GGIR and calculate metrics, structures the milestone data in epoch level data
}
\examples{
# Example usage:
# load.structure.axivity(tz = "Europe/Amsterdam", filepath = "/path/to/accelerometer/data", 
#                         outputdir = "/path/to/save/milestone/data", validhours = 12, epochlength = 5, 
#                         processeddir = "/path/to/save/processed/data", overwrit = FALSE)

}
\references{
The GGIR package: https://cran.r-project.org/package=GGIR
}
\keyword{GGIR}
\keyword{accelerometer,}
\keyword{data}
\keyword{processing,}
