% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.nonwear.R
\name{remove.nonwear}
\alias{remove.nonwear}
\title{remove.nonwear}
\usage{
remove.nonwear(epochdir.hip, epochdir.wrist, savefolder, data.castor)
}
\arguments{
\item{epochdir.hip}{Path to the root directory containing hip accelerometer data aggregated in epochs.}

\item{epochdir.wrist}{Path to the root directory containing wrist accelerometer data aggregated in epochs.}

\item{savefolder}{Subdirectory where the preprocessed valid data will be stored.}

\item{data.castor}{Data frame containing information on accelerometer attachment date and time and wear protocol deviations}
}
\description{
'remove.nonwear' removes timestamp data indicated as non-wear from accelerometer data.
}
