% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.json.R
\name{unnest.json}
\alias{unnest.json}
\title{unnest.json}
\usage{
unnest.json(datadir, date)
}
\arguments{
\item{datadir}{A character string specifying the directory where the .json files are located.}

\item{date}{A character string specifying the date to be used in the output filenames.}
}
\value{
The function does not return any values. It saves two CSV files: one for motor milestones and one for activities.
}
\description{
'unnest.json' reads .json files from a specified directory, processes the data, and saves the results as CSV files. The .json files required are information about participants, intake questions, registered activities, and questionnaires.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads JSON data files: participants.json, measurementperiods.json, measuringmoments.json, and questionnaires.json.
\item Converts the JSON files to data frames, processing and unnesting the nested JSON structure.
\item Links research codes with participant data.
\item Saves the processed data as CSV files in the specified directory.
}
}
