% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validday.app.R
\name{validday.app}
\alias{validday.app}
\title{validday.app}
\usage{
validday.app(data, minhours = 12, mindays = 7, savedir, CoDa = TRUE)
}
\arguments{
\item{data}{A data.frame object that contains the app data}

\item{minhours}{An integer indicating the minimal recorded hours/day for a day to be considered a valid day (default = 12).}

\item{mindays}{An integer that indicates the minimal number of days to be considered valid data (default = 7).}

\item{savedir}{Path to the directory where the preprocessed valid data will be stored.}

\item{CoDa}{A boolean indicating if the compositional scores (isometric log ratios) need to be calculated (default = TRUE)}
}
\value{
A data.frame containing the valid preprocessed data based on the specified criteria.
}
\description{
'validday.app' preprocesses the app data with the activity duration per category per day by removing invalid days
}
\details{
This function preprocesses app data, removing days with insufficient activity duration per day.
It filters data based on the given minimum recorded hours/day and minimum number of valid days.
The resulting valid data will be saved in the specified directory.
}
\examples{
# Example usage:
# validday.app(data = my_data, minhours = 12, mindays = 5, savedir = "path/to/save")

}
