% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validday.axivity.R
\name{validday.axivity}
\alias{validday.axivity}
\title{validday.axivity}
\usage{
validday.axivity(
  epochdir,
  epochlength = 5,
  minhours = 12,
  mindays = 2,
  savedir,
  castordata
)
}
\arguments{
\item{epochdir}{Path to the root directory containing epoch-level accelerometer data in .RData files.}

\item{epochlength}{Integer indicating the epoch length in seconds, default is 5.}

\item{minhours}{Integer indicating the minimum recorded hours per day for a day to be considered valid, default is 12.}

\item{mindays}{Integer indicating the minimum number of valid days required, default is 2.}

\item{savedir}{Path to the directory where the preprocessed valid data will be stored.}

\item{castordata}{Data frame containing information about measurement periods and participant IDs.}
}
\description{
'validday.axivity' preprocesses epoch-level accelerometer data by removing days that do not meet validity criteria.
}
